feat - Renamed project and added new classes and functionality.
