--- PredicateTests.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30309.148
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PredicateTests", "PredicateTests\PredicateTests.csproj", "{F633CCCB-0D95-4165-962A-D885BCB624A9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F633CCCB-0D95-4165-962A-D885BCB624A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F633CCCB-0D95-4165-962A-D885BCB624A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F633CCCB-0D95-4165-962A-D885BCB624A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F633CCCB-0D95-4165-962A-D885BCB624A9}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D6A67252-83A5-497E-8B31-F77A19296546}
-	EndGlobalSection
-EndGlobal
