+++ Src/PredicateTests/IEntity.cs
@@ -1,0 +1,3 @@
+﻿namespace POCPredicate;
+
+public interface IEntity { }
