+++ Src/PredicateTests/Teste.cs
@@ -1,0 +1,10 @@
+﻿namespace POCPredicate;
+
+public class Teste : IEntity
+{
+    public bool IsActive { get; set; }
+
+    public string Document { get; set; }
+
+    public string Name { get; set; }
+}
